+++ lib/events_api/doings.ex
@@ -13,11 +13,10 @@ defmodule EventsApi.Doings do
   ## Examples
 
       iex> create_event(%{field: value})
-      {:ok, %User{}}
+      {:ok, %Event{}}
 
       iex> create_event(%{field: bad_value})
       {:error, %Ecto.Changeset{}}
-
   """
   def create_event(attrs \\ %{}) do
     %Event{}
@@ -25,11 +24,37 @@ defmodule EventsApi.Doings do
     |> Repo.insert()
   end
 
+  @doc """
+  Função lista todos os eventos persistidos no banco de dados.
+
+  ## Examples
+
+      iex> list_events()
+      [%Event{}, ...]
+  """
   def list_events do
     Repo.all(Event)
   end
 
+  @doc """
+  Função responsavel por retornar um evento pelo id.
+
+  ## Paramentros da função
+  - id: número de identificação do evento
+
+  ## Informações adcionais
+    - caso o id passado esteja errado ou não exista é retornado um erro `{:error, :not_found}`.
+
+  ## Example empty lists
+
+       iex> get_event!("a37a802c-2466-41c6-b84e-15396694f886")
+       {:ok, %Event{}}
+  """
   def get_event!(id) do
-    Repo.get!(Event, id)
+    Repo.get_by(Event, id: id)
+    |> case do
+      nil -> {:error, :not_found}
+      event -> {:ok, event}
+    end
   end
 end