+++ test/events_api/accounts_test.exs
@@ -51,7 +51,9 @@ defmodule EventsApi.AccountsTest do
     test "retornando um usuário pelo id" do
       {:ok, user} = Accounts.create_user(@valid_params)
 
-      assert Accounts.get_user!(user.id).name == "Henry"
+      {:ok, result} = Accounts.get_user!(user.id)
+
+      assert result.name == "Henry"
     end
   end
 end