+++ test/events_api/doings_test.exs
@@ -39,8 +39,9 @@ defmodule EventsApi.DoingsTest do
       user = user_fixture()
       params = Map.put(@valid_event, :user_id, user.id)
       {:ok, event} = Doings.create_event(params)
+      {:ok, result} = Doings.get_event!(event.id)
 
-      assert Doings.get_event!(event.id).title == "Inteligência Emocional"
+      assert result.title == "Inteligência Emocional"
     end
   end
 end