+++ test/events_api_web/views/user_view_test.exs
@@ -17,4 +17,23 @@ defmodule EventsApiWeb.UserViewTest do
              email: "some@email.com"
            } = UserView.render("show.json", %{user: user})
   end
+
+  test "render/2 returns auth_user" do
+    user = %{
+      id: "1",
+      name: "some displayName",
+      email: "some@email.com",
+      password: "123123"
+    }
+
+    token = "ABCDEFJH"
+
+    assert %{
+             user: %{
+               name: "some displayName",
+               email: "some@email.com",
+               token: "ABCDEFJH"
+             }
+           } = UserView.render("user_auth.json", %{user: user, token: token})
+  end
 end