+++ CHANGELOG.md
@@ -14,3 +14,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuração do travis ci para garantir o processo de integração contínua
 - configuração codecov para relatório de teste
 - configuração de CD
+
+## [0.0.2] - 2021-02-11
+### Added
+- rotas para criar e exibir usuários
+- rotas para criar e exibir eventos
+- controller de usuários e eventos
+- criação de regra de negócio para eventos
+- teste para controller de eventos e usuários
+- teste para regras de negócios
+- add lib guardian para autenticação
+
+## [0.0.1] - 2021-02-12
+### Added
+- criação de modulos do guardian: Guardian, Pipeline e error_handler
+- Adição de rota e função signin no controller do usuário
+- pipeline de atenticação
+- usuário para criar um evento e listar o usuário por id tem que está logado
+- teste de autenticação
+- teste de token
+- teste de signin
+- escrita da documentação
+- escrita do readme
+